Implement full version checking logic.
I've added a short list of MRI versions that are currently supported.
Hardcoding this list is obviously not a great idea as I'd have to
update the gem at irregular intervals, and I'd probably forget to do
so when it matters most. I'll see if ruby-lang.org is interested in
creating a small endpoint that returns currently supported versions.
Same with JRuby, Rubinius, and other implementations.

Something like ruby-lang.org/versions.json which would return:
[
  "supported: [
    "2.0.0"
  ],
  "unsupported": [
    "1.8.7",
    "1.9.2",
    "1.9.3"
  ]
]

The current output is not very nice when the current directory contains
a lot of sub-directories so I'll have to see how to improve that. For
directories with few sub-directories however it looks alright:

$ liability -r  .
Doing a recursive Ruby liability check on '.' ...
Doing a simple Ruby liability check on 'elf' ...
Warning: The following Ruby versions are used inside of 'elf':
1.9.2